+++ crates/core/src/potentials/mod.rs
@@ -5,8 +5,11 @@ pub mod pair;
 
 use serde::{Deserialize, Serialize};
 
-use crate::potentials::coulomb::{CoulombMeta, CoulombPotential};
-use crate::potentials::pair::{PairMeta, PairPotential};
+use crate::internal::Float;
+use crate::neighbors::{NeighborList, Neighbors, NeighborsBuilder};
+use crate::potentials::pair::PairPotential;
+use crate::system::species::Specie;
+use crate::system::System;
 
 /// Base trait for all potentials.
 #[typetag::serde(tag = "type")]
@@ -15,60 +18,68 @@ pub trait Potential: Send + Sync {}
 /// Container type to hold instances of each potential in the system.
 #[derive(Serialize, Deserialize)]
 pub struct Potentials {
-    pub coulombs: Vec<(CoulombMeta, Box<dyn CoulombPotential>)>,
-    pub pairs: Vec<(PairMeta, Box<dyn PairPotential>)>,
+    neighbors: Neighbors,
+    pairs: Vec<Box<dyn PairPotential>>,
+    pub update_interval: usize,
+}
+
+impl Potentials {
+    pub fn setup(&mut self, system: &System) {
+        self.neighbors.setup(system);
+    }
+
+    pub fn update(&mut self, system: &System) {
+        self.neighbors.update(system);
+    }
+
+    pub fn pair_interactions(&self) -> Vec<(&dyn PairPotential, usize, usize)> {
+        self.neighbors
+            .iter()
+            .map(move |((i, j), index)| (&*self.pairs[*index], *i, *j))
+            .collect()
+    }
 }
 
 /// Constructor for the [`Potentials`](velvet_core::potentials::Potentials) type.
 pub struct PotentialsBuilder {
-    coulombs: Vec<(CoulombMeta, Box<dyn CoulombPotential>)>,
-    pairs: Vec<(PairMeta, Box<dyn PairPotential>)>,
+    neighbors_builder: NeighborsBuilder,
+    pairs: Vec<Box<dyn PairPotential>>,
+    update_interval: usize,
 }
 
 impl PotentialsBuilder {
     /// Returns a new `PotentialsBuilder`.
     pub fn new() -> PotentialsBuilder {
         PotentialsBuilder {
-            coulombs: Vec::new(),
+            neighbors_builder: NeighborsBuilder::new(),
             pairs: Vec::new(),
+            update_interval: 1,
         }
     }
 
-    /// Adds a new coulomb potential to the collection.
-    ///
-    /// # Arguments
-    ///
-    /// * `meta` - Coulomb metadata.
-    /// * `potential` - Boxed coulomb potential.
-    pub fn add_coulomb(
+    pub fn with_pair(
         mut self,
-        meta: CoulombMeta,
-        potential: Box<dyn CoulombPotential>,
-    ) -> PotentialsBuilder {
-        self.coulombs.push((meta, potential));
-        self
-    }
-
-    /// Adds a new pair potential to the collection.
-    ///
-    /// # Arguments
-    ///
-    /// * `meta` - Pairwise metadata.
-    /// * `potential` - Boxed pair potential.
-    pub fn add_pair(
-        mut self,
-        meta: PairMeta,
         potential: Box<dyn PairPotential>,
+        cutoff: Float,
+        species: (Specie, Specie),
     ) -> PotentialsBuilder {
-        self.pairs.push((meta, potential));
+        let nl = NeighborList::new(cutoff, Some(species));
+        self.neighbors_builder = self.neighbors_builder.with_neighbor_list(nl);
+        self.pairs.push(potential);
         self
     }
 
-    /// Returns an initialized `Potentials`.
+    pub fn with_update_interval(mut self, interval: usize) -> PotentialsBuilder {
+        self.update_interval = interval;
+        self
+    }
+
     pub fn build(self) -> Potentials {
+        let neighbors = self.neighbors_builder.build();
         Potentials {
-            coulombs: self.coulombs,
+            neighbors,
             pairs: self.pairs,
+            update_interval: self.update_interval,
         }
     }
 }