+++ crates/core/src/properties/energy.rs
@@ -5,33 +5,6 @@ use crate::potentials::Potentials;
 use crate::properties::{IntrinsicProperty, Property};
 use crate::system::System;
 
-/// Potential energy due to coulombic potentials.
-#[derive(Clone, Copy, Debug, Serialize, Deserialize)]
-pub struct CoulombEnergy;
-
-impl Property for CoulombEnergy {
-    type Res = Float;
-
-    fn calculate(&self, system: &System, potentials: &Potentials) -> Self::Res {
-        let mut energy = 0.0;
-        for (meta, potential) in &potentials.coulombs {
-            for (i, j) in &meta.indices {
-                let elem_i = system.elements[*i];
-                let elem_j = system.elements[*j];
-                let pos_i = system.positions[*i];
-                let pos_j = system.positions[*j];
-                let r = system.cell().distance(&pos_i, &pos_j);
-                if meta.cutoff > r {
-                    energy += potential.energy(elem_i.charge(), elem_j.charge(), r)
-                        - potential.energy_self(elem_i.charge())
-                        - potential.energy_self(elem_j.charge());
-                }
-            }
-        }
-        energy
-    }
-}
-
 /// Potential energy due to pairwise potentials.
 #[derive(Clone, Copy, Debug, Serialize, Deserialize)]
 pub struct PairEnergy;
@@ -41,24 +14,13 @@ impl Property for PairEnergy {
 
     fn calculate(&self, system: &System, potentials: &Potentials) -> Self::Res {
         potentials
-            .pairs
+            .pair_interactions()
             .iter()
-            .map(|(meta, potential)| {
-                let energy: Float = meta
-                    .indices
-                    .iter()
-                    .map(|(i, j)| {
-                        let pos_i = system.positions[*i];
-                        let pos_j = system.positions[*j];
-                        let r = system.cell().distance(&pos_i, &pos_j);
-                        if r < meta.cutoff {
-                            potential.energy(r)
-                        } else {
-                            0.0
-                        }
-                    })
-                    .sum();
-                energy
+            .map(|(potential, i, j)| {
+                let pos_i = system.positions[*i];
+                let pos_j = system.positions[*j];
+                let r = system.cell.distance(&pos_i, &pos_j);
+                potential.energy(r)
             })
             .sum()
     }
@@ -72,9 +34,8 @@ impl Property for PotentialEnergy {
     type Res = Float;
 
     fn calculate(&self, system: &System, potentials: &Potentials) -> Self::Res {
-        let coulomb_energy = CoulombEnergy.calculate(system, potentials);
         let pair_energy = PairEnergy.calculate(system, potentials);
-        coulomb_energy + pair_energy
+        pair_energy
     }
 }
 
@@ -87,10 +48,13 @@ impl IntrinsicProperty for KineticEnergy {
 
     fn calculate_intrinsic(&self, system: &System) -> <Self as IntrinsicProperty>::Res {
         let kinetic_energy: Float = system
-            .elements
+            .specie_ids
             .iter()
             .zip(system.velocities.iter())
-            .map(|(elem, vel)| 0.5 * elem.mass() * vel.norm_squared())
+            .map(|(id, vel)| {
+                let sp = system.species[id];
+                0.5 * sp.mass() * vel.norm_squared()
+            })
             .sum();
         kinetic_energy
     }