+++ crates/core/src/properties/temperature.rs
@@ -17,7 +17,7 @@ impl IntrinsicProperty for Temperature {
         let kinetic = KineticEnergy.calculate_intrinsic(system);
         // NOTE: Calculating DOF this way is a potentially nasty bug if future
         // support is added for degrees of freedom beyond just 3D particles.
-        let dof = (system.size() * 3) as Float;
+        let dof = (system.size * 3) as Float;
         2.0 * kinetic / (dof * BOLTZMANN)
     }
 }