+++ crates/test-utils/src/lib.rs
@@ -1,10 +1,10 @@
 use std::fs::File;
 use std::io::BufReader;
 
-use velvet_core::potentials::coulomb::{CoulombMeta, Wolf};
-use velvet_core::potentials::pair::{LennardJones, PairMeta};
+use velvet_core::potentials::pair::LennardJones;
 use velvet_core::potentials::{Potentials, PotentialsBuilder};
 use velvet_core::system::elements::Element;
+use velvet_core::system::species::Specie;
 use velvet_core::system::System;
 use velvet_external_data::poscar::load_poscar;
 
@@ -32,50 +32,49 @@ pub fn xenon_system() -> System {
     load_poscar(reader)
 }
 
-pub fn argon_potentials(system: &System) -> Potentials {
+pub fn argon_potentials() -> Potentials {
+    let argon = Specie::from_element(0, Element::Ar);
     let lj = LennardJones::new(4.184, 3.4);
-    let meta = PairMeta::new(8.5, (Element::Ar, Element::Ar), system);
     PotentialsBuilder::new()
-        .add_pair(meta, Box::new(lj))
+        .with_pair(Box::new(lj), 8.5, (argon, argon))
         .build()
 }
 
-pub fn binary_gas_potentials(system: &System) -> Potentials {
+pub fn binary_gas_potentials() -> Potentials {
+    let argon = Specie::from_element(0, Element::Ar);
+    let xenon = Specie::from_element(1, Element::Xe);
     let lj0 = LennardJones::new(4.184, 3.4);
-    let meta0 = PairMeta::new(12.0, (Element::Ar, Element::Ar), &system);
     let lj1 = LennardJones::new(7.824, 4.57);
-    let meta1 = PairMeta::new(12.0, (Element::Xe, Element::Xe), &system);
     let lj2 = LennardJones::new(6.276, 4.0);
-    let meta2 = PairMeta::new(12.0, (Element::Ar, Element::Xe), &system);
     PotentialsBuilder::new()
-        .add_pair(meta0, Box::new(lj0))
-        .add_pair(meta1, Box::new(lj1))
-        .add_pair(meta2, Box::new(lj2))
+        .with_pair(Box::new(lj0), 12.0, (argon, argon))
+        .with_pair(Box::new(lj1), 12.0, (xenon, xenon))
+        .with_pair(Box::new(lj2), 12.0, (argon, xenon))
         .build()
 }
 
-pub fn rocksalt_potentials(system: &System) -> Potentials {
-    let wolf = Wolf::new(8.0);
-    let wolf_meta = CoulombMeta::new(8.0, &system);
-    let lj0 = LennardJones::new(0.00598, 4.612);
-    let lj0_meta = PairMeta::new(8.0, (Element::Cl, Element::Cl), &system);
-    let lj1 = LennardJones::new(0.0187, 2.497);
-    let lj1_meta = PairMeta::new(8.0, (Element::Na, Element::Na), &system);
-    let lj2 = LennardJones::new(0.01058, 3.5545);
-    let lj2_meta = PairMeta::new(8.0, (Element::Cl, Element::Na), &system);
-    PotentialsBuilder::new()
-        .add_coulomb(wolf_meta, Box::new(wolf))
-        .add_pair(lj0_meta, Box::new(lj0))
-        .add_pair(lj1_meta, Box::new(lj1))
-        .add_pair(lj2_meta, Box::new(lj2))
-        .build()
-}
+// pub fn rocksalt_potentials(system: &System) -> Potentials {
+//     let wolf = Wolf::new(8.0);
+//     let wolf_meta = CoulombMeta::new(8.0, &system);
+//     let lj0 = LennardJones::new(0.00598, 4.612);
+//     let lj0_meta = PairMeta::new(8.0, (Element::Cl, Element::Cl), &system);
+//     let lj1 = LennardJones::new(0.0187, 2.497);
+//     let lj1_meta = PairMeta::new(8.0, (Element::Na, Element::Na), &system);
+//     let lj2 = LennardJones::new(0.01058, 3.5545);
+//     let lj2_meta = PairMeta::new(8.0, (Element::Cl, Element::Na), &system);
+//     PotentialsBuilder::new()
+//         .add_coulomb(wolf_meta, Box::new(wolf))
+//         .add_pair(lj0_meta, Box::new(lj0))
+//         .add_pair(lj1_meta, Box::new(lj1))
+//         .add_pair(lj2_meta, Box::new(lj2))
+//         .build()
+// }
 
-pub fn xenon_potentials(system: &System) -> Potentials {
+pub fn xenon_potentials() -> Potentials {
+    let xenon = Specie::from_element(0, Element::Xe);
     let lj = LennardJones::new(7.824, 4.57);
-    let meta = PairMeta::new(12.0, (Element::Xe, Element::Xe), system);
     PotentialsBuilder::new()
-        .add_pair(meta, Box::new(lj))
+        .with_pair(Box::new(lj), 12.0, (xenon, xenon))
         .build()
 }